[show][muxcable] add support for show mux firmware version all

+++ show/muxcable.py
@@ -97,6 +97,45 @@ def check_port_in_mux_cable_table(port):
     return False
 
 
+
+def get_per_port_firmware(port):
+
+    state_db = {}
+    muxcable_info_tbl = {}
+    mux_info_dict = {}
+
+    # Getting all front asic namespace and correspding config and state DB connector
+
+    namespaces = multi_asic.get_front_end_namespaces()
+    for namespace in namespaces:
+        asic_id = multi_asic.get_asic_index_from_namespace(namespace)
+        state_db[asic_id] = db_connect("STATE_DB", namespace)
+        muxcable_info_tbl[asic_id] = swsscommon.Table(state_db[asic_id], "MUX_CABLE_INFO")
+
+    if platform_sfputil is not None:
+        asic_index = platform_sfputil_helper.get_asic_id_for_logical_port(port)
+
+    if asic_index is None:
+        # TODO this import is only for unit test purposes, and should be removed once sonic_platform_base
+        # is fully mocked
+        import sonic_platform_base.sonic_sfp.sfputilhelper
+        asic_index = sonic_platform_base.sonic_sfp.sfputilhelper.SfpUtilHelper().get_asic_id_for_logical_port(port)
+        if asic_index is None:
+            click.echo("Got invalid asic index for port {}, cant retrieve mux cable table entries".format(port))
+            return False
+
+    (status, fvp) = muxcable_info_tbl[asic_index].get(port)
+    if status is False : 
+        click.echo("Got invalid status for state DB, cant retrieve mux cable info entries".format(port))
+        return False
+
+    res_dir = dict(fvp)
+    mux_info_dict["version_nic_active"] = res_dir.get("version_nic_active", None)
+    mux_info_dict["version_peer_active"] = res_dir.get("version_peer_active", None)
+    mux_info_dict["version_self_active"] = res_dir.get("version_self_active", None)
+
+    return mux_info_dict
+
 def get_response_for_version(port, mux_info_dict):
     state_db = {}
     xcvrd_show_fw_res_tbl = {}
@@ -1528,7 +1567,7 @@ def version(db, port, active):
     delete_all_keys_in_db_table("STATE_DB", "XCVRD_SHOW_FW_RSP")
     delete_all_keys_in_db_table("STATE_DB", "XCVRD_SHOW_FW_RES")
 
-    if port is not None:
+    if port is not None and port != "all":
 
         res_dict = {}
         mux_info_dict, mux_info_active_dict = {}, {}
@@ -1561,6 +1600,60 @@ def version(db, port, active):
             click.echo("{}".format(json.dumps(mux_info_active_dict, indent=4)))
         else:
             click.echo("{}".format(json.dumps(mux_info_dict, indent=4)))
+
+   elif port == "all" and port is not None:
+
+        logical_port_list = platform_sfputil_helper.get_logical_list()
+
+        rc_exit = True
+        print_data = []
+
+        for port in logical_port_list:
+
+            if platform_sfputil is not None:
+                physical_port_list = platform_sfputil_helper.logical_port_name_to_physical_port_list(port)
+
+            if not isinstance(physical_port_list, list):
+                continue
+            if len(physical_port_list) != 1:
+                continue
+
+            if not check_port_in_mux_cable_table(port):
+                continue
+
+            physical_port = physical_port_list[0]
+           
+            logical_port_list_for_physical_port = platform_sfputil_helper.get_physical_to_logical()
+
+            logical_port_list_per_port = logical_port_list_for_physical_port.get(physical_port, None)
+
+            """ This check is required for checking whether or not this logical port is the one which is
+            actually mapped to physical port and by convention it is always the first port.
+            TODO: this should be removed with more logic to check which logical port maps to actual physical port
+            being used"""
+
+            if port != logical_port_list_per_port[0]:
+                continue
+
+
+            port = platform_sfputil_helper.get_interface_alias(port, db)
+            
+            mux_info = get_per_port_firmware(port)
+            if not isinstance(mux_info, dict):
+                mux_info = {}
+            
+            for key, val in mux_info.items():
+                print_port_data = []
+                port = platform_sfputil_helper.get_interface_alias(port, db)
+                print_port_data.append(port)
+                print_port_data.append(key)
+                print_port_data.append(val)
+                print_data.append(print_port_data)
+
+        headers = ['PORT', 'SIDE', 'VERSION']
+        click.echo(tabulate(print_data, headers=headers))
+
+        sys.exit(CONFIG_SUCCESSFUL) 
     else:
         port_name = platform_sfputil_helper.get_interface_name(port, db)
         click.echo("Did not get a valid Port for mux firmware version".format(port_name))
@@ -1640,6 +1733,7 @@ def metrics(db, port, json_output):
 def event_log(db, port, json_output):
     """Show muxcable event log <port>"""
 
+    click.confirm(('Muxcable at port {} will retreive cable logs from MCU, approx time could be ~2 minutes wait time Continue?'.format(port), abort=True)
     port = platform_sfputil_helper.get_interface_name(port, db)
     delete_all_keys_in_db_table("APPL_DB", "XCVRD_EVENT_LOG_CMD")
     delete_all_keys_in_db_table("STATE_DB", "XCVRD_EVENT_LOG_RSP")
@@ -1735,7 +1829,7 @@ def packetloss(db, port, json_output):
     for namespace in namespaces:
         asic_id = multi_asic.get_asic_index_from_namespace(namespace)
 
-        per_npu_statedb[asic_id] = swsscommon.SonicV2Connector(use_unix_socket_path=True, namespace=namespace)
+        per_npu_statedb[asic_id] = swsscommon.SonicV2Connector(use_unix_socket_path=False, namespace=namespace)
         per_npu_statedb[asic_id].connect(per_npu_statedb[asic_id].STATE_DB)
 
         pckloss_table_keys[asic_id] = per_npu_statedb[asic_id].keys(
